:construction: Refactor daily activity to common extension (#601)
