+++ teferi/Extensions/TimeSlotServiceExtensions.swift
@@ -1,0 +1,28 @@
+import Foundation
+
+extension TimeSlotService
+{
+    func getActivities(forDate date: Date) -> [Activity]
+    {
+        let result = getTimeSlots(forDay: date)
+                        .groupBy(category)
+                        .map(toActivity)
+        
+        return result
+    }
+    
+    private func category(of timeSlot: TimeSlot) -> Category
+    {
+        return timeSlot.category
+    }
+    
+    private func toActivity(_ timeSlots: [TimeSlot]) -> Activity
+    {
+        let totalTime =
+            timeSlots
+                .map(calculateDuration)
+                .reduce(0, +)
+        
+        return Activity(category: timeSlots.first!.category, duration: totalTime)
+    }
+}
